all: add awwan scripts to deploy and test image
The gcloud-image-publish.aww is the script to publish the recently
build image to gcloud and create new image.

The gcloud-image-test.aww is the script to test the new image by
creating new compute engine.

The gcloud-test-image-official.aww is the script to test the official
public image from Google.

The gcloud-test-tail.aww is the script to get the serial log from
test image in compute engine.