all: use worldwide pacman mirrorlist
We have an issue where the image build on local and the generated
mirrorlist is located in India (or Pakistan?).
This cause our package installation/update really slow when creating a
VM in Singapore region.

Since we did not know who or where the image to be used, the safe
configuration should be to use the worldwide one.