all: update README
Remove list of choices that are not applicable again on the latest build
script.

Mention list of Linux Guest Environment packages that we install on the
image, including the link to its AUR repository.

Update the steps and sample commands for "Build your own image" to use
gcloud instead of gsutil.